Warning fixes (everywhere)

+++ 32blit-pico/input.cpp
@@ -73,13 +73,13 @@ void update_input() {
   using namespace blit;
 
 #ifdef INPUT_GPIO
-  api.buttons = (get_button(ButtonIO::LEFT) ? Button::DPAD_LEFT : 0)
-              | (get_button(ButtonIO::RIGHT) ? Button::DPAD_RIGHT : 0)
-              | (get_button(ButtonIO::UP) ? Button::DPAD_UP : 0)
-              | (get_button(ButtonIO::DOWN) ? Button::DPAD_DOWN : 0)
-              | (get_button(ButtonIO::A) ? Button::A : 0)
-              | (get_button(ButtonIO::B) ? Button::B : 0)
-              | (get_button(ButtonIO::X) ? Button::X : 0)
-              | (get_button(ButtonIO::Y) ? Button::Y : 0);
+  api.buttons = (get_button(ButtonIO::LEFT)  ? uint32_t(Button::DPAD_LEFT) : 0)
+              | (get_button(ButtonIO::RIGHT) ? uint32_t(Button::DPAD_RIGHT) : 0)
+              | (get_button(ButtonIO::UP)    ? uint32_t(Button::DPAD_UP) : 0)
+              | (get_button(ButtonIO::DOWN)  ? uint32_t(Button::DPAD_DOWN) : 0)
+              | (get_button(ButtonIO::A)     ? uint32_t(Button::A) : 0)
+              | (get_button(ButtonIO::B)     ? uint32_t(Button::B) : 0)
+              | (get_button(ButtonIO::X)     ? uint32_t(Button::X) : 0)
+              | (get_button(ButtonIO::Y)     ? uint32_t(Button::Y) : 0);
 #endif
 }
